% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life-expect.R
\name{clean_life}
\alias{clean_life}
\title{Clean life expectancy data}
\usage{
clean_life(path)
}
\arguments{
\item{path}{A path to a directory files.}
}
\value{
A data frame.
}
\description{
Clean life expectancy data
}
