% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties-package.R
\docType{package}
\name{counties-package}
\alias{counties}
\alias{counties-package}
\title{counties: US county data}
\description{
US county data by year for demographics, income, and health.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rwright88/counties}
  \item Report bugs at \url{https://github.com/rwright88/counties/issues}
}

}
\author{
\strong{Maintainer}: Ray Wright \email{rwright88@gmail.com}

}
\keyword{internal}
