% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demographics}
\alias{demographics}
\title{Demographics}
\format{A data frame with seven variables:
\describe{
\item{\code{year}}{Year}
\item{\code{county_fips}}{FIPS county code}
\item{\code{population}}{Population estimate as of July 1}
\item{\code{asian}}{Distribution of population that is Asian}
\item{\code{black}}{Distribution of population that is Black}
\item{\code{hispanic}}{Distribution of population that is Hispanic}
\item{\code{white}}{Distribution of population that is White}
}

For further details, see \url{https://www2.census.gov/programs-surveys/popest/datasets/}}
\usage{
demographics
}
\description{
Demographics and population data by year and county from the US Census
Bureau. Population is the population estimate as of July 1 of that year.
}
\keyword{datasets}
