% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-income.R
\name{get_income}
\alias{get_income}
\title{Get income data}
\usage{
get_income()
}
\value{
\code{tbl_df} with seven columns: \code{year}, \code{income_hh}
}
\description{
Get income data by year and county from the US Census Bureau. Income is
median household income in nominal dollars.
}
\examples{
\dontrun{
get_pop()
}
}
\references{
\url{https://www2.census.gov/programs-surveys/saipe/datasets/}
}
