% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-life.R
\name{get_life}
\alias{get_life}
\title{Get life expectancy data}
\usage{
get_life()
}
\value{
\code{tbl_df} with three columns: \code{year}, \code{county_fips},
        \code{life_expect}
}
\description{
Get life expectancy data by year and county from the Global Health Data Exchange
}
\examples{
\dontrun{
get_life()
}
}
\references{
\url{http://ghdx.healthdata.org/us-data}
}
