% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pop.R
\name{get_pop}
\alias{get_pop}
\title{Get population data}
\usage{
get_pop()
}
\value{
\code{tbl_df} with seven columns: \code{year}, \code{county_fips},
        \code{population}, \code{asian}, \code{black}, \code{hispa},
        \code{white}
}
\description{
Get population and demograhpic data by year and county from the US Census
Bureau. Population is the population estimate as of July 1 of that year.
}
\examples{
\dontrun{
get_pop()
}
}
\references{
\url{https://www2.census.gov/programs-surveys/popest/datasets/}
}
