% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life}
\alias{life}
\title{Life expectancy}
\format{A data frame with three variables:
\describe{
\item{\code{year}}{Year}
\item{\code{county_fips}}{FIPS county code}
\item{\code{life_expect}}{Life expectancy at birth}
}

For further details, see \url{http://ghdx.healthdata.org/us-data}}
\usage{
life
}
\description{
life expectancy data by year and county from the Global Health Data Exchange.
}
\keyword{datasets}
