% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-income.R
\name{read_income}
\alias{read_income}
\title{Read income data.}
\usage{
read_income(path)
}
\arguments{
\item{path}{A path to a file.}
}
\value{
\code{tbl_df} with 3 columns: \code{state}, \code{county},
        \code{income_hh}, \code{year}
}
\description{
Read income data.
}
