% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life-expect.R
\name{read_life}
\alias{read_life}
\title{Read life expectancy data.}
\usage{
read_life(path)
}
\arguments{
\item{path}{A path to a file.}
}
\value{
A data frame.
}
\description{
Read life expectancy data.
}
