% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pop.R
\name{read_pop}
\alias{read_pop}
\title{Read population data.}
\usage{
read_pop(path)
}
\arguments{
\item{path}{A path to a file.}
}
\value{
\code{tbl_df} with 18 columns: \code{STATE}, \code{COUNTY},
        \code{YEAR}, \code{AGEGRP}, \code{NHWA_MALE}, \code{NHWA_MALE},
        \code{NHWA_FEMALE}, \code{NHBA_MALE}, \code{NHBA_FEMALE},
        \code{NHIA_MALE}, \code{NHIA_FEMALE}, \code{NHAA_MALE},
        \code{NHAA_FEMALE}, \code{NHNA_MALE}, \code{NHNA_FEMALE},
        \code{NHTOM_MALE}, \code{NHTOM_FEMALE}, \code{H_MALE},
        \code{H_FEMALE}
}
\description{
Read population data.
}
