% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pop.R
\name{rec_race}
\alias{rec_race}
\title{Recode race variables from population data.}
\usage{
rec_race(x)
}
\arguments{
\item{x}{Race variables.}
}
\value{
A vector the same size as \code{x}.
}
\description{
Recode race variables from population data.
}
