% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life-expect.R
\name{get_life}
\alias{get_life}
\title{Get life expectancy data}
\usage{
get_life(path)
}
\arguments{
\item{path}{A path to a directory of files.}
}
\value{
A data frame.
}
\description{
Get life expectancy data
}
