% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income}
\alias{income}
\title{Income}
\format{A data frame with three variables:
\describe{
\item{\code{county_fips}}{FIPS county code}
\item{\code{year}}{Year}
\item{\code{income_hh}}{Median household income}
}

For further details, see \url{https://www2.census.gov/programs-surveys/saipe/datasets/}}
\usage{
income
}
\description{
Income data by county and year from the US Census Bureau. Income is median
household income in nominal dollars.
}
\keyword{datasets}
